fields: improved money field to always have a mask
MoneyField is poorly usable without a mask. This commit adds a default
custom mask, relying on a modified version of jquery-mask-plugin and
Django's jQuery. The advantage of this approach is that it works out
of the box on both normal Django forms and Admin Django forms.

There's still some features to be implemented, like currency mask,
customizable currency and apply the mask to dynamically loaded fields